requirements.txt
Flask
Flask-SQLAlchemy
flask-bcrypt
python-dotenv